# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CommentResponse:
    properties:
      body:
        type: string
      comment_votes:
        items:
          $ref: '#/definitions/models.CommentVote'
        type: array
      created_at:
        type: string
      downvotes:
        type: integer
      id:
        type: string
      net_votes:
        description: upvotes - downvotes
        type: integer
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      sentiment_label:
        type: string
      sentiment_score:
        type: number
      upvotes:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
      user_vote:
        description: '"up", "down", or null'
        type: string
    type: object
  handlers.CommentsListResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/handlers.CommentResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  handlers.CreateCommentRequest:
    properties:
      body:
        example: Great product! Highly recommended.
        maxLength: 2000
        minLength: 1
        type: string
      product_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    required:
    - body
    - product_id
    type: object
...
  handlers.StandardErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
...
  handlers.UpdateCommentRequest:
    properties:
      body:
        example: Updated comment text
        maxLength: 2000
        minLength: 1
        type: string
    required:
    - body
    type: object
...
  handlers.VoteCommentRequest:
    properties:
      vote_type:
        enum:
        - up
        - down
        example: up
        type: string
    required:
    - vote_type
    type: object
  models.Category:
    properties:
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
  models.Comment:
    properties:
      body:
        type: string
      comment_votes:
        items:
          $ref: '#/definitions/models.CommentVote'
        type: array
      created_at:
        type: string
      downvotes:
        type: integer
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      sentiment_label:
        type: string
      sentiment_score:
        type: number
      upvotes:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.CommentVote:
    properties:
      comment:
        $ref: '#/definitions/models.Comment'
      comment_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
      vote_type:
        type: string
    type: object
  models.Favorite:
    properties:
      favorited_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
...
  models.Product:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/models.User'
      curated_price:
        type: number
      curated_tags:
        items:
          type: string
        type: array
      description:
        type: string
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      feature_vector:
        $ref: '#/definitions/models.ProductFeatureVector'
      id:
        type: string
      name:
        type: string
      price:
        type: number
      product_suggestions:
        items:
          $ref: '#/definitions/models.ProductSuggestion'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      similarity_data_as_product:
        items:
          $ref: '#/definitions/models.ProductSimilarityData'
        type: array
      similarity_data_as_similar:
        items:
          $ref: '#/definitions/models.ProductSimilarityData'
        type: array
      tags:
        items:
          type: string
        type: array
      user_events:
        items:
          $ref: '#/definitions/models.UserEvent'
        type: array
    type: object
  models.ProductFeatureVector:
    properties:
      embedding:
        items:
          type: number
        type: array
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.ProductSuggestion:
    properties:
      generated_at:
        type: string
      id:
        type: string
      model_version:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      reason:
        type: string
      suggested_price_max:
        type: number
      suggested_price_min:
        type: number
      suggested_tags:
        items:
          type: string
        type: array
    type: object
...
  models.User:
    properties:
      birth_year:
        type: integer
      comment_votes:
        items:
          $ref: '#/definitions/models.CommentVote'
        type: array
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      email:
        type: string
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      http_request_logs:
        items:
          $ref: '#/definitions/models.HttpRequestLog'
        type: array
      id:
        type: string
      is_admin:
        type: boolean
      name:
        type: string
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      region:
        type: string
      user_events:
        items:
          $ref: '#/definitions/models.UserEvent'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@bachelor-ecommerce.com
    name: API Support
    url: http://www.swagger.io/support
  description: ML-Powered E-commerce Platform API with advanced recommendation system,
    intelligent search, and comprehensive analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bachelor E-commerce API
  version: "1.0"
paths:
...
  /comments:
    post:
      consumes:
      - application/json
      description: Create a new comment on a product
      parameters:
      - description: Comment creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Comment created successfully
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a comment
      tags:
      - Comments
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a comment (only by the author or admin)
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Comment deleted successfully
        "400":
          description: Invalid comment ID
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "403":
          description: User not authorized to delete this comment
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update an existing comment (only by the author)
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Comment updated successfully
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "403":
          description: User not authorized to update this comment
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Update comment
      tags:
      - Comments
  /comments/{id}/vote:
    post:
      consumes:
      - application/json
      description: Vote up or down on a comment
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Vote data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.VoteCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Vote recorded successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "409":
          description: User already voted on this comment
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Vote on comment
      tags:
      - Comments
  /comments/my:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of comments by a specific user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comments retrieved successfully
          schema:
            $ref: '#/definitions/handlers.CommentsListResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's comments
      tags:
      - Comments
  /comments/product/{product_id}:
    get:
      consumes:
      - application/json
      description: Get paginated comments for a specific product with optional user
        vote information
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: '"created_at"'
        description: Sort by (created_at, upvotes, downvotes)
        in: query
        name: sort_by
        type: string
      - default: '"desc"'
        description: Sort order (asc, desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comments retrieved successfully
          schema:
            $ref: '#/definitions/handlers.CommentsListResponse'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      summary: Get product comments
      tags:
      - Comments
  /comments/stats:
    get:
      consumes:
      - application/json
      description: Get statistics about comments (for admin or analytics)
      produces:
      - application/json
      responses:
        "200":
          description: Comment statistics
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Get comment statistics
      tags:
      - Comments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...